geofabrik_zones {geofabrik}	R Documentation

sf data frame containing attributes and geometries of geofrabric download zones

Description

sf data frame containing attributes and geometries of geofrabric download zones

Note

The 'geofabrik_zones' can be (re-)generated using the script in the 'data-raw' directory.

[Package geofabrik version 0.1.0 Index]