get_geofabrik {geofabrik}	R Documentation

Download OSM data from [geofabrik.de](http://www.geofabrik.de/)

Description

Download OSM data from [geofabrik.de](http://www.geofabrik.de/)

Usage

get_geofabrik(
  name = "west-yorkshire",
  layer = "lines",
  ...,
  attributes = make_additional_attributes(layer = layer),
  download_directory = gf_download_directory(),
  ask = TRUE,
  max_dist = 5,
  op = sf::st_contains
)

Arguments

name
String or ['sf::sf'] spatial object of the geofabrik zone to download. See examples.

layer
Character string telling 'sf' which OSM layer to import. One of 'points', 'lines' (the default), 'multilinestrings', 'multipolygons' or 'other_relations'

...
Additional arguments passed to ['read_pbf()']

attributes
Vector of character strings naming attributes to import

download_directory
Where to download the data? 'tempdir()' by default. If you want to download your data into a persistent directory, set 'GF_DOWNLOAD_DIRECTORY=/path/for/osm/data' in your '.Renviron' file, e.g. with 'usethis::edit_r_environ()'.

ask
Should the user be asked before downloading the file?

max_dist
What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download?

op
The binary spatial predicate used to identify the smallest geofabrik zones that matches the simple feature input in 'name'

Examples

get_geofabrik("isle of man")
andorra = get_geofabrik(name = "andorra") # try other names, e.g. name = "west-yorkshire"
head(andorra)
cycleways_andorra = get_geofabrik("andorra", key = "highway", value = "cycleway")
plot(cycleways_andorra)
# user asked to choose closest match when interactive
# get_geofabrik("kdljfdl", ask = FALSE) # not run to save time
# get zone associated with a point
name = sf::st_sfc(sf::st_point(c(-1.3, 50.7)), crs = 4326)
get_geofabrik(name)
name = sf::st_sfc(sf::st_point(c(0, 53)), sf::st_point(c(-2, 55)), crs = 4326)
gf_find_sf(name)

[Package geofabrik version 0.1.0 Index]