gf_filename {geofabrik}	R Documentation

Get the filename of a file downloaded from geofabrik

Description

Get the filename of a file downloaded from geofabrik

Usage

gf_filename(
  name = "West Yorkshire",
  download_directory = gf_download_directory()
)

Arguments

name
The name of the geofrabic zone. Must be an element in 'geofabrik::geofabrik_zones$name'.

download_directory
Where to download the data? 'tempdir()' by default. If you want to download your data into a persistent directory, set 'GF_DOWNLOAD_DIRECTORY=/path/for/osm/data' in your '.Renviron' file, e.g. with 'usethis::edit_r_environ()'.

Value

A character vector.

Examples

f = gf_filename("Isle of Wight")
f
file.exists(f)

[Package geofabrik version 0.1.0 Index]