gf_find_sf {geofabrik}	R Documentation

Find geofabrik zones based on sf or sfc object

Description

Find geofabrik zones based on sf or sfc object

Usage

gf_find_sf(name, ask = FALSE, op = sf::st_contains)

Arguments

name
String or ['sf::sf'] spatial object of the geofabrik zone to download. See examples.

ask
Should the user be asked before downloading the file?

op
The binary spatial predicate used to identify the smallest geofabrik zones that matches the simple feature input in 'name'

Value

A data frame representing the matching items from the Geofabrik website

Examples

name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
gf_find_sf(name)

[Package geofabrik version 0.1.0 Index]