make_additional_attributes {geofabrik}	R Documentation

Get modified version of config file for reading .pbf files with GDAL/sf

Description

This function provides a user friendly interface to the [GDAL OSM driver](https://gdal.org/drivers/vector/osm.html) via 'sf'. See https://github.com/r-spatial/sf/issues/1157 for details.

Usage

make_additional_attributes(layer)

make_ini_attributes(
  attributes,
  layer,
  defaults = get_ini_layer_defaults(layer),
  append = TRUE
)

Arguments

layer
Which layer to read in? One of "points" "lines" "multipolygons" "multilinestrings" or "other_relations"

attributes
Vector of character strings naming attributes to import

defaults
The default attributes

append
Should the columns named in 'attributes' be appended to the default columns? 'TRUE' by default.

Details

Default additional attributes were taken from https://taginfo.openstreetmap.org/keys

Value

A character vector representing an ini file.

Examples

make_ini_attributes("oneway", "lines")

[Package geofabrik version 0.1.0 Index]